Fix minor error
-Added check for message history for deleting the bot's earlier message

+++ FFBEBot/src/util/Selector.java
@@ -179,7 +179,9 @@ public class Selector {
 	 */
 	private static void selfPrune(Select s,MessageReceivedEvent e){
 		try{
-			e.getChannel().getMessageById(s.messageID).complete().delete().queue();
+			if(e.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_HISTORY)){
+				e.getChannel().getMessageById(s.messageID).complete().delete().queue();
+			}
 			if(e.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_MANAGE)){
 				e.getMessage().delete().queue();
 				Log.log("DELETE", "deleted message "+e.getMessage().getContentRaw());
